LEIA: Avoid throwing an instance of 'std::length_error'
When calling just "LEIA" (!= ./LEIA), argv[0] is just "LEIA" (unlike $0
in shell), so the old code was not finding / and then raising an
exception.

There's no easy and portable way to find the path to the executable (or:
if you have one, feel free to add it to the code), so let's just use
hardcoded constants when we can't find the file.